# contracts/openapi/content-v1.yaml
openapi: 3.1.0
info:
  title: GTU Blog Content API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /public/posts:
    get:
      operationId: listPublicPosts
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Published posts ordered by first publication time and id, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostSummaryPage' }
  /public/posts/{slug}:
    get:
      operationId: getPublicPost
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Published post detail with citations linked to the post revisions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostDetailResponse' }
        '404': { description: Published post not found. }
  /public/search:
    get:
      operationId: searchPublicPosts
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Published posts matching full-text content or category/tag names.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostSummaryPage' }
  /public/categories/{slug}:
    get:
      operationId: listPublicPostsByCategory
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Published posts in a category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostSummaryPage' }
  /public/tags/{slug}:
    get:
      operationId: listPublicPostsByTag
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Published posts in a tag.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostSummaryPage' }
  /public/archive:
    get:
      operationId: listPublicArchiveMonths
      responses:
        '200':
          description: Published post counts by archive month.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PostArchiveEntryResponse' }
  /public/archive/{year}/{month}:
    get:
      operationId: listPublicArchiveMonthPosts
      parameters:
        - name: year
          in: path
          required: true
          schema: { type: integer, minimum: 1000, maximum: 9998 }
        - name: month
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 12 }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Published posts whose first publication timestamp falls within the UTC calendar month.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostSummaryPage' }
  /admin/posts:
    get:
      operationId: listAdminPosts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Active administrator post list ordered by creation time and id, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostSummaryPage' }
  /admin/posts/stats:
    get:
      operationId: getAdminPostStats
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Administrator post counts by lifecycle state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostStatsResponse' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
  schemas:
    PostSummaryPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PostSummaryResponse' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 50 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    PostSummaryResponse:
      type: object
      required: [id, slug, title, excerpt, status, viewCount, categories, tags, categoryDetails, tagDetails]
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string }
        title: { type: string }
        excerpt: { type: string }
        status: { $ref: '#/components/schemas/PostStatus' }
        firstPublishedAt:
          type: [string, 'null']
          description: Offset-less ISO-8601 local date-time serialized from Java LocalDateTime; null before publication.
        viewCount: { type: integer, format: int64, minimum: 0 }
        categories:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        categoryDetails:
          type: array
          items: { $ref: '#/components/schemas/TaxonomyItemResponse' }
        tagDetails:
          type: array
          items: { $ref: '#/components/schemas/TaxonomyItemResponse' }
    PostDetailResponse:
      type: object
      required: [id, slug, title, excerpt, contentMarkdown, contentHtml, status, createdAt, updatedAt, viewCount, categories, tags, citations, relatedPosts]
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string }
        title: { type: string }
        excerpt: { type: string }
        contentMarkdown: { type: string }
        contentHtml: { type: string }
        status: { $ref: '#/components/schemas/PostStatus' }
        firstPublishedAt:
          type: [string, 'null']
          description: Offset-less ISO-8601 local date-time serialized from Java LocalDateTime; null before publication.
        createdAt:
          type: string
          description: Offset-less ISO-8601 local date-time serialized from Java LocalDateTime.
        updatedAt:
          type: string
          description: Offset-less ISO-8601 local date-time serialized from Java LocalDateTime.
        viewCount: { type: integer, format: int64, minimum: 0 }
        categories:
          type: array
          items: { $ref: '#/components/schemas/TaxonomyItemResponse' }
        tags:
          type: array
          items: { $ref: '#/components/schemas/TaxonomyItemResponse' }
        citations:
          type: array
          items: { $ref: '#/components/schemas/PostCitationResponse' }
        relatedPosts:
          type: array
          items: { $ref: '#/components/schemas/PostSummaryResponse' }
    PostCitationResponse:
      type: object
      required: [url, title]
      properties:
        url: { type: string, format: uri }
        title: { type: [string, 'null'] }
    PostStatsResponse:
      type: object
      required: [total, published, draft, archived, deleted]
      properties:
        total: { type: integer, format: int64, minimum: 0 }
        published: { type: integer, format: int64, minimum: 0 }
        draft: { type: integer, format: int64, minimum: 0 }
        archived: { type: integer, format: int64, minimum: 0 }
        deleted: { type: integer, format: int64, minimum: 0 }
    TaxonomyItemResponse:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
    PostArchiveEntryResponse:
      type: object
      required: [year, month, count]
      properties:
        year: { type: integer, minimum: 1000, maximum: 9998 }
        month: { type: integer, minimum: 1, maximum: 12 }
        count: { type: integer, format: int64, minimum: 0 }
    PostStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED, DELETED]